fix delete and add calculation functionality

+++ src/routes/(dashboard)/invoices/LineItemRow.svelte
@@ -1,33 +1,69 @@
 <script lang="ts">
   import { createEventDispatcher } from 'svelte';
   import Trash from '$lib/components/Icon/Trash.svelte';
+  import { twoDecimals, dollarsToCents } from '$lib/utils/moneyHelers';
 
   export let lineItem: LineItem;
+  export let canDelete: boolean = false;
+
   let dispatch = createEventDispatcher();
+  let unitPrice: string = twoDecimals(lineItem.amount / lineItem.quantity);
+  let amount = twoDecimals(lineItem.amount);
+  // chamge values on the fly reactivly
+  $: {
+    amount = twoDecimals(lineItem.quantity * Number(unitPrice));
+    lineItem.amount = dollarsToCents(Number(amount));
+  }
 </script>
 
 <!-- Structure -->
 <div class="invoice-line-item border-b-2 border-fog py-2">
   <div>
-    <input class="line-item" type="text" name="description" />
+    <input class="line-item" type="text" name="description" bind:value={lineItem.description} />
   </div>
   <div>
-    <input class="line-item text-right" type="number" name="unitPrice" step="0.01" min="0" />
+    <input
+      class="line-item text-right"
+      type="number"
+      name="unitPrice"
+      step="0.01"
+      min="0"
+      bind:value={unitPrice}
+      on:blur={() => {
+        unitPrice = twoDecimals(Number(unitPrice));
+      }}
+    />
   </div>
   <div>
-    <input class="line-item text-center" type="number" name="quantity" min="0" />
+    <input
+      class="line-item text-right"
+      type="number"
+      name="quantity"
+      min="0"
+      bind:value={lineItem.quantity}
+    />
   </div>
   <div>
-    <input class="line-item text-right" type="number" name="amount" step="0.01" min="0" />
+    <input
+      class="line-item text-right"
+      type="number"
+      name="amount"
+      step="0.01"
+      min="0"
+      bind:value={amount}
+      disabled
+    />
   </div>
   <div>
-    <button
-      class="center h-10 w-10 text-pastelPurple hover:text-lavenderIndigo"
-      type="button"
-      on:click|preventDefault={() => dispatch('removeLineItem', lineItem.id)}
-    >
-      <Trash />
-    </button>
+    {#if canDelete}
+      <button
+        class="center h-10 w-10 text-pastelPurple hover:text-lavenderIndigo"
+        type="button"
+        on:click|preventDefault={() => dispatch('removeLineItem', lineItem.id)}
+      >
+        <Trash />
+      </button>
+    {/if}
   </div>
 </div>
 
@@ -45,5 +81,9 @@
   input[type='number']:focus,
   input[type='text']:focus {
     @apply border-solid border-lavenderIndigo outline-none;
+  }
+  input[type='text']:disabled,
+  input[type='number']:disabled {
+    @apply border-b-0 bg-transparent px-0;
   }
 </style>
